ai-pipeline/automation-orchestrator.py
#!/usr/bin/env python3
"""
MILAPP - Orquestrador de Automação IA-IA
Integração automática entre Lovable IA e Cursor IA
"""

import asyncio
import json
import logging
from datetime import datetime
from pathlib import Path
from typing import Dict, List, Optional
import aiohttp
import yaml

class AIIntegrationOrchestrator:
    def __init__(self, config_path: str = "ai-pipeline/config.yaml"):
        self.config = self.load_config(config_path)
        self.lovable_client = LovableAIClient(self.config['lovable'])
        self.cursor_client = CursorAIClient(self.config['cursor'])
        self.git_client = GitAutomationClient(self.config['git'])
        self.deployment_client = DeploymentClient(self.config['deployment'])
        
        # Estado do pipeline
        self.current_task = None
        self.task_history = []
        self.approval_queue = []
        
        logging.basicConfig(level=logging.INFO)
        self.logger = logging.getLogger(__name__)
    
    def load_config(self, config_path: str) -> Dict:
        """Carrega configuração do pipeline"""
        with open(config_path, 'r') as f:
            return yaml.safe_load(f)
    
    async def start_automation_pipeline(self, task_description: str):
        """Inicia pipeline de automação IA-IA"""
        self.logger.info(f"🚀 Iniciando pipeline para: {task_description}")
        
        # 1. Análise inicial com Cursor IA
        analysis = await self.cursor_client.analyze_requirements(task_description)
        
        # 2. Geração com Lovable IA
        generated_code = await self.lovable_client.generate_implementation(analysis)
        
        # 3. Revisão e refinamento com Cursor IA
        refined_code = await self.cursor_client.review_and_refine(generated_code)
        
        # 4. Testes automáticos
        test_results = await self.run_automated_tests(refined_code)
        
        # 5. Análise de qualidade
        quality_report = await self.cursor_client.quality_analysis(refined_code)
        
        # 6. Preparar para aprovação
        approval_package = self.prepare_approval_package(
            task_description, refined_code, test_results, quality_report
        )
        
        self.approval_queue.append(approval_package)
        
        self.logger.info("✅ Pipeline concluído - Aguardando aprovação")
        return approval_package
    
    async def run_automated_tests(self, code_changes: Dict) -> Dict:
        """Executa testes automatizados"""
        test_results = {
            'unit_tests': await self.run_unit_tests(code_changes),
            'integration_tests': await self.run_integration_tests(code_changes),
            'security_scan': await self.run_security_scan(code_changes),
            'performance_tests': await self.run_performance_tests(code_changes)
        }
        return test_results
    
    def prepare_approval_package(self, task: str, code: Dict, tests: Dict, quality: Dict) -> Dict:
        """Prepara pacote para aprovação"""
        return {
            'id': f"task_{datetime.now().strftime('%Y%m%d_%H%M%S')}",
            'task_description': task,
            'code_changes': code,
            'test_results': tests,
            'quality_report': quality,
            'summary': self.generate_summary(code, tests, quality),
            'risk_assessment': self.assess_risks(code, tests, quality),
            'created_at': datetime.now().isoformat(),
            'status': 'pending_approval'
        }
    
    async def approve_changes(self, task_id: str, approved: bool, comments: str = ""):
        """Aprova ou rejeita mudanças"""
        task = next((t for t in self.approval_queue if t['id'] == task_id), None)
        if not task:
            raise ValueError(f"Task {task_id} não encontrada")
        
        if approved:
            # Aplicar mudanças
            await self.apply_changes(task['code_changes'])
            
            # Commit e push automático
            await self.git_client.commit_and_push(
                task['code_changes'], 
                f"feat: {task['task_description']} - Aprovado via IA-IA Pipeline"
            )
            
            # Deploy automático (se configurado)
            if self.config['deployment']['auto_deploy']:
                await self.deployment_client.deploy(task['code_changes'])
            
            self.logger.info(f"✅ Mudanças aprovadas e aplicadas: {task_id}")
        else:
            # Feedback para melhorias
            await self.process_rejection(task, comments)
            self.logger.info(f"❌ Mudanças rejeitadas: {task_id}")
        
        task['status'] = 'approved' if approved else 'rejected'
        task['approval_comments'] = comments
        task['approved_at'] = datetime.now().isoformat()
    
    async def apply_changes(self, code_changes: Dict):
        """Aplica mudanças no código"""
        for file_path, content in code_changes['files'].items():
            file_path = Path(file_path)
            file_path.parent.mkdir(parents=True, exist_ok=True)
            
            with open(file_path, 'w', encoding='utf-8') as f:
                f.write(content)
    
    def generate_summary(self, code: Dict, tests: Dict, quality: Dict) -> str:
        """Gera resumo das mudanças"""
        return f"""
📋 **Resumo das Mudanças**
- Arquivos modificados: {len(code.get('files', {}))}
- Testes passaram: {tests.get('unit_tests', {}).get('passed', 0)}/{tests.get('unit_tests', {}).get('total', 0)}
- Qualidade: {quality.get('overall_score', 0)}/100
- Complexidade: {quality.get('complexity', 'Baixa')}
        """.strip()
    
    def assess_risks(self, code: Dict, tests: Dict, quality: Dict) -> Dict:
        """Avalia riscos das mudanças"""
        risks = []
        
        if quality.get('overall_score', 0) < 70:
            risks.append("Qualidade do código abaixo do padrão")
        
        if tests.get('security_scan', {}).get('vulnerabilities', 0) > 0:
            risks.append("Vulnerabilidades de segurança detectadas")
        
        if tests.get('unit_tests', {}).get('passed', 0) < tests.get('unit_tests', {}).get('total', 0):
            risks.append("Testes unitários falharam")
        
        return {
            'risk_level': 'Alto' if len(risks) > 2 else 'Médio' if len(risks) > 0 else 'Baixo',
            'risks': risks,
            'recommendations': self.generate_recommendations(risks)
        }
    
    def generate_recommendations(self, risks: List[str]) -> List[str]:
        """Gera recomendações baseadas nos riscos"""
        recommendations = []
        
        if "Qualidade do código abaixo do padrão" in risks:
            recommendations.append("Revisar padrões de código antes de aprovar")
        
        if "Vulnerabilidades de segurança detectadas" in risks:
            recommendations.append("Corrigir vulnerabilidades antes do deploy")
        
        if "Testes unitários falharam" in risks:
            recommendations.append("Corrigir testes antes de prosseguir")
        
        return recommendations

class LovableAIClient:
    """Cliente para integração com Lovable IA"""
    
    def __init__(self, config: Dict):
        self.api_key = config['api_key']
        self.base_url = config['base_url']
    
    async def generate_implementation(self, analysis: Dict) -> Dict:
        """Gera implementação baseada na análise do Cursor IA"""
        prompt = self.build_lovable_prompt(analysis)
        
        # Simulação da chamada para Lovable IA
        # Em produção, seria uma chamada real para a API
        return {
            'files': {
                'src/components/NewFeature.tsx': '// Código gerado pelo Lovable IA',
                'src/hooks/useNewFeature.ts': '// Hook gerado pelo Lovable IA',
                'tests/NewFeature.test.tsx': '// Testes gerados pelo Lovable IA'
            },
            'metadata': {
                'generated_by': 'Lovable IA',
                'prompt_used': prompt,
                'confidence_score': 0.85
            }
        }
    
    def build_lovable_prompt(self, analysis: Dict) -> str:
        """Constrói prompt para Lovable IA baseado na análise do Cursor"""
        return f"""
        Com base na análise do Cursor IA, gere a implementação para:
        
        Contexto: {analysis.get('context', '')}
        Requisitos: {analysis.get('requirements', [])}
        Arquitetura: {analysis.get('architecture', {})}
        Padrões: {analysis.get('patterns', [])}
        
        Gere código seguindo os padrões do projeto MILAPP.
        """

class CursorAIClient:
    """Cliente para integração com Cursor IA"""
    
    def __init__(self, config: Dict):
        self.api_key = config['api_key']
        self.base_url = config['base_url']
    
    async def analyze_requirements(self, task_description: str) -> Dict:
        """Analisa requisitos usando Cursor IA"""
        # Simulação da análise
        return {
            'context': 'Análise do contexto do projeto',
            'requirements': ['req1', 'req2', 'req3'],
            'architecture': {'frontend': 'React', 'backend': 'FastAPI'},
            'patterns': ['Component Pattern', 'Hook Pattern'],
            'dependencies': ['react', 'fastapi'],
            'estimated_complexity': 'Média'
        }
    
    async def review_and_refine(self, generated_code: Dict) -> Dict:
        """Revisa e refina código gerado pelo Lovable IA"""
        # Simulação da revisão
        refined_code = generated_code.copy()
        refined_code['metadata']['refined_by'] = 'Cursor IA'
        refined_code['metadata']['improvements'] = [
            'Otimização de performance',
            'Melhoria na tipagem TypeScript',
            'Adição de tratamento de erros'
        ]
        return refined_code
    
    async def quality_analysis(self, code: Dict) -> Dict:
        """Análise de qualidade do código"""
        return {
            'overall_score': 85,
            'complexity': 'Média',
            'maintainability': 'Alta',
            'test_coverage': 90,
            'security_score': 95,
            'performance_score': 88
        }

class GitAutomationClient:
    """Cliente para automação Git"""
    
    def __init__(self, config: Dict):
        self.repo_path = config['repo_path']
        self.branch = config['branch']
    
    async def commit_and_push(self, changes: Dict, message: str):
        """Commit e push automático"""
        # Implementação real seria com gitpython ou subprocess
        print(f"Git: Commitando mudanças - {message}")

class DeploymentClient:
    """Cliente para deploy automático"""
    
    def __init__(self, config: Dict):
        self.deployment_url = config['url']
        self.api_key = config['api_key']
    
    async def deploy(self, changes: Dict):
        """Deploy automático"""
        # Implementação real seria com chamada para API de deploy
        print(f"Deploy: Aplicando mudanças automaticamente")

# Configuração do pipeline
CONFIG = {
    'lovable': {
        'api_key': 'your_lovable_api_key',
        'base_url': 'https://api.lovable.com'
    },
    'cursor': {
        'api_key': 'your_cursor_api_key',
        'base_url': 'https://api.cursor.com'
    },
    'git': {
        'repo_path': '.',
        'branch': 'main'
    },
    'deployment': {
        'auto_deploy': True,
        'url': 'https://deploy.milapp.com',
        'api_key': 'your_deploy_api_key'
    }
}

# Exemplo de uso
async def main():
    orchestrator = AIIntegrationOrchestrator()
    
    # Iniciar pipeline
    task = "Criar nova funcionalidade de dashboard de métricas de automação"
    approval_package = await orchestrator.start_automation_pipeline(task)
    
    # Simular aprovação
    await orchestrator.approve_changes(approval_package['id'], True, "Aprovado!")

if __name__ == "__main__":
    asyncio.run(main()) 