ai-pipeline/requirements.txt
# MILAPP - Requirements para Pipeline IA-IA
# Integração automática entre Lovable IA e Cursor IA

# Core dependencies
aiohttp==3.9.1
asyncio==3.4.3
pyyaml==6.0.1
python-dotenv==1.0.0

# Dashboard
streamlit==1.29.0
pandas==2.1.4
plotly==5.17.0

# Git automation
gitpython==3.1.40

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Security scanning
bandit==1.7.5
safety==2.3.5

# Performance testing
locust==2.17.0

# Monitoring
prometheus-client==0.19.0

# Caching
redis==5.0.1

# Logging
structlog==23.2.0

# Encryption
cryptography==41.0.8

# HTTP client
httpx==0.25.2

# JSON handling
jsonschema==4.20.0

# Date/time
python-dateutil==2.8.2

# File handling
pathlib2==2.3.7

# Configuration
click==8.1.7
rich==13.7.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.8.0 